增加其它角色离开视线的逻辑
1. unit管理组件改到GameManager中
2. 服务端unit改为以RoleId创建
