--- examples/Communication-MMO-Examples/Server/Hotfix/Handler/C2M_StopMoveMessageHandler.cs
@@ -1,10 +1,0 @@
-﻿namespace Fantasy;
-
-public class C2M_StopMoveMessageHandler : Addressable<Unit, C2M_StopMoveMessage>
-{
-    protected override async FTask Run(Unit unit, C2M_StopMoveMessage message)
-    {
-        MoveHelper.Stop(unit, NoticeClientType.Aoi);
-        await FTask.CompletedTask;
-    }
-}
