--- examples/Communication-MMO-Examples/Unity/Assets/MMORes/Resources/Prefabs/Unit/Follower.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 44b7238fc8844ac47a473b2f10086bf7
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
