--- examples/Communication-MMO-Examples/Unity/Assets/Scenes/PreBall-Test.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1bfaffdd5da82ec4d8a633924b92dc0d
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
