+++ examples/Communication-MMO-Examples/Unity/Assets/Scripts/MMOGame/ModelViewer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be80d40ac05522443a866f3b046a6193
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
