+++ examples/Communication-MMO-Examples/Unity/Assets/Scripts/MMOGame/ModelViewer/ClassViewer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4d331e6516681846a8ad25d7e173f83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
