--- examples/Communication-MMO-Examples/Unity/Assets/Scripts/MMOGame/Viewer/UnitViewer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 69e5bfb5fca86014f9b657aeb54f6376
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
