+++ examples/Communication-MMO-Examples/Unity/Assets/Scripts/MicroFramework/Models.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c8a5a236edb86c44a0aa289f45637e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
