+++ examples/Communication-MMO-Examples/Unity/Assets/Scripts/MicroFramework/Models/BaseViewer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc60693933416544fafb610db2f1d810
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
