+++ examples/Communication-MMO-Examples/Unity/Assets/Scripts/MicroFramework/Models/GameFacade.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a3fd82ee1122b1545b310cebd5bb0915
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
