+++ examples/Communication-MMO-Examples/Unity/Assets/Scripts/Network/Handler/M2C_UnitHelperHandler.cs
@@ -29,7 +29,7 @@ namespace BestGame
 	{
 		protected override async FTask Run(Session session, M2C_UnitRemove message)
 		{
-			
+			Sender.Ins.Recive(ReciveType.RemvoeUnits,message);
 			await FTask.CompletedTask;
 		}
 	}
